remove need for herocontext so that metadata can be added to each top level page
